feat: implement schema version toggle functionality in TemplateSchemaMessageRenderer and update GenuiTemplate components for improved schema management
